validation/xx084xx--dynstat__agents-mcp-clients/pyproject.toml
[project]
name = "agents-mcp-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.12.1",
    "mcp[cli]>=1.7.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.8",
]
